Update deck.md
Expanded significantly on the explanation of time needed for research with different-rarity cards under the "What should I research next?" question. Added an additional example to help illustrate how research time is based on the rarity of the resulting card, and tweaked the existing example to illustrate an exception to the basic "same rarity" advice that was previously given.