Update compare.html to improve layout
Setting max-height on the tbody when it's wider than the content area means that on non-touch devices, you have to scroll all the way to the right in #results before you can access the vertical scrollbar of tbody. Removed tbdoy's max-height and set it on the #results fieldset instead. Changed from static 460px to 100vh (100% of viewport height).
I also removed max-width from .content to take advantage of available space on larger displays, since the table is often more cards wide than will fit in 800px - nobody wants to have to scroll horizontally. As this feature is currently designed, it may not always be possible to avoid horizontal scrolling, but now a large screen will avoid it when possible, or at least minimize it otherwise, rather than just having a large swath of whitespace on both sides.