Fix for Issue #7
It looks like the site owner will need to edit the skill_targetIP.png that I provided by extending the image's background to the right and down in order to double both dimensions (to add whitespace without resizing the icon itself) in order to make it display properly. (I would do this myself, but the only image editing program I have is MSPaint, which does not preserve PNG transparency.)
Then the site owner will need to upload the image to their repository with the name "skill_targetIP.png", and then this code change should correctly display the new image on "only targets show" skills.